+++ pkg/scheduler/load_profile.go
@@ -3,6 +3,6 @@ package scheduler
 import "time"
 
 type LoadProfile interface {
-	At(elapsed time.Duration) int
+	At(elapsed time.Duration) uint64
 	TotalDuration() time.Duration
 }
